chore(runtime): initial deletion of shadow emulation code
