+++ packages/runtime/src/dom.ts
@@ -1,5 +1,4 @@
 import { DI, IContainer, IResolver } from '@aurelia/kernel';
-import { ICustomElement } from './templating/custom-element';
 
 export interface INodeLike {
   readonly firstChild: INode | null;
@@ -63,24 +62,15 @@ const childObserverOptions = { childList: true };
 function returnTrue(): true {
   return true;
 }
+
 function returnFalse(): false {
   return false;
 }
-function createElementViewHostWithShadowDOM(node: Element, options?: ShadowRootInit): INode {
-  if (options) {
-    return node.attachShadow(options);
-  }
 
-  (<any>node).$usingSlotEmulation = true;
-  return node;
-}
-function createElementViewHostWithNoShadowDOM(node: Element): INode {
-  (<any>node).$usingSlotEmulation = true;
-  return node;
-}
 function removeNormal(node: Element): void {
   node.remove();
 }
+
 function removePolyfilled(node: Element): void {
   // not sure if we still actually need this, this used to be an IE9/10 thing
   node.parentNode.removeChild(node);
@@ -97,51 +87,17 @@ export const DOM = {
   createElement(name: string): INode {
     return document.createElement(name);
   },
+
   createAnchor(): INode {
     return document.createComment('anchor');
   },
+
   /*@internal*/createTemplate(): INode {
     return document.createElement('template');
   },
 
-  createChildObserver(parent: INode, onChildrenChanged: MutationCallback, options?: MutationObserverInit): IChildObserver {
-    if ((<any>parent).$usingSlotEmulation) {
-      return (<any>parent).$component.$contentView.attachChildObserver(onChildrenChanged);
-    } else {
-      const observer = new MutationObserver(onChildrenChanged);
-      (<any>observer).childNodes = parent.childNodes;
-      observer.observe(<Node>parent, options || childObserverOptions);
-      return <any>observer;
-    }
-  },
-
-  platformSupportsShadowDOM(): boolean {
-    if (HTMLElement.prototype.attachShadow === undefined) {
-      return (DOM.platformSupportsShadowDOM = returnFalse)();
-    } else {
-      return (DOM.platformSupportsShadowDOM = returnTrue)();
-    }
-  },
-
-  createElementViewHost(node: INode, options?: ShadowRootInit): INode {
-    // only check shadowDOM availability once and then set a short-circuited function directly to save a few cycles
-    if (DOM.platformSupportsShadowDOM()) {
-      return (DOM.createElementViewHost = createElementViewHostWithShadowDOM)(<Element>node, options);
-    } else {
-      return (DOM.createElementViewHost = createElementViewHostWithNoShadowDOM)(<Element>node);
-    }
-  },
-
   cloneNode(node: INode, deep?: boolean): INode {
     return (<Node>node).cloneNode(deep !== false); // use true unless the caller explicitly passes in false
-  },
-
-  getCustomElementForNode(node: INode): ICustomElement | null {
-    return (<any>node).$component || null; //Set during component $hydrate
-  },
-
-  isUsingSlotEmulation(node: INode): boolean {
-    return (<any>node).$usingSlotEmulation === true;
   },
 
   isNodeInstance(potentialNode: any): potentialNode is INode {
@@ -220,7 +176,7 @@ export const DOM = {
     if ((<any>node).auInterpolationTarget === true) {
       return false;
     }
-    const text = (<any>node).textContent;
+    const text = (node as Node).textContent;
     const len = text.length;
     let i = 0;
     // for perf benchmark of this compared to the regex method: http://jsben.ch/p70q2 (also a general case against using regex)
