+++ packages/runtime/src/templating/custom-element.ts
@@ -8,7 +8,6 @@ import { AttachLifecycle, DetachLifecycle, IAttach, IBindSelf } from './lifecycl
 import { IRenderSlot } from './render-slot';
 import { IRenderingEngine } from './rendering-engine';
 import { IRuntimeBehavior } from './runtime-behavior';
-import { ShadowDOMEmulation } from './shadow-dom';
 import { IContentView, IViewOwner, View } from './view';
 
 export interface ICustomElementType extends IResourceType<ITemplateSource, ICustomElement> { }
@@ -112,10 +111,12 @@ export const CustomElementResource : IResourceKind<ITemplateSource, ICustomEleme
 
       this.$context = template.renderContext;
       this.$behavior = renderingEngine.applyRuntimeBehavior(Type, this, description.bindables);
+
       this.$host = description.containerless ? DOM.convertToAnchor(host, true) : host;
       this.$shadowRoot = DOM.createElementViewHost(this.$host, description.shadowOptions);
       this.$usingSlotEmulation = DOM.isUsingSlotEmulation(this.$host);
       this.$contentView = View.fromCompiledContent(this.$host, options);
+
       this.$view = this.$behavior.hasCreateView
         ? (this as any).createView(host, options.parts, template)
         : template.createFor(this, host, options.parts);
@@ -174,12 +175,6 @@ export const CustomElementResource : IResourceKind<ITemplateSource, ICustomEleme
 
       if (this.$slot !== null) {
         this.$slot.$attach(encapsulationSource, lifecycle);
-      }
-
-      //Native ShadowDOM would be distributed as soon as we append the view below.
-      //So, we emulate the distribution of nodes at the same time.
-      if (this.$contentView !== null && this.$slots) {
-        ShadowDOMEmulation.distributeContent(this.$contentView, this.$slots);
       }
 
       if (description.containerless) {
@@ -265,4 +260,122 @@ export function createCustomElementDescription(templateSource: ITemplateSource, 
     shadowOptions: templateSource.shadowOptions || (Type as any).shadowOptions || null,
     hasSlots: templateSource.hasSlots || false
   };
+}
+
+
+
+
+export interface IContentView extends IView {
+  childObserver?: IChildObserver;
+  insertVisualChildBefore(visual: IVisual, refNode: INode);
+  removeVisualChild(visual: IVisual);
+  attachChildObserver(onChildrenChanged: () => void): IChildObserver;
+}
+
+/*@internal*/
+export interface IContentViewChildObserver extends IChildObserver {
+  notifyChildrenChanged(): void;
+}
+
+// This is the default and only implementation of IContentView.
+// This type of view represents the content that is placed between the opening and closing tags
+// of a custom element. It is only used when ShadowDOM Emulation is required. It enables
+// various interactions between the content, as a logical view, the ShadowDom Emulation,
+// and the RenderSlot.
+// Most if not all of this will go away when we remove the shadow dom emulation and render slot.
+// However, we may need to keep the IChildObserver abstraction for enabling the $children property
+// in different scenarios, depending on what optimizations we make around when/when not to use shadow dom.
+/*@internal*/
+export class ContentView implements IContentView {
+  public firstChild: INode;
+  public lastChild: INode;
+  public childNodes: INode[];
+  public childObserver: IContentViewChildObserver = null;
+
+  constructor(private contentHost: INode) {
+    const childNodes = this.childNodes = Array.from(contentHost.childNodes)
+    this.firstChild = childNodes[0];
+    this.lastChild = childNodes[childNodes.length - 1];
+  }
+
+  public attachChildObserver(onChildrenChanged: () => void): IChildObserver {
+    const contentViewNodes = this.childNodes;
+    let observer = this.childObserver;
+
+    if (!observer) {
+      this.childObserver = observer = {
+        get childNodes() {
+          return contentViewNodes;
+        },
+        disconnect() {
+          onChildrenChanged = null;
+        },
+        notifyChildrenChanged() {
+          if (onChildrenChanged !== null) {
+            onChildrenChanged();
+          }
+        }
+      };
+
+      const workQueue = Array.from(contentViewNodes);
+
+      while(workQueue.length) {
+        const current = workQueue.shift();
+
+        if ((current as any).$isContentProjectionSource) {
+          const contentIndex = contentViewNodes.indexOf(current);
+
+          ((current as any).$slot as IRenderSlot).children.forEach(x => {
+            const childNodes = x.$view.childNodes;
+            childNodes.forEach(x => workQueue.push(x));
+            contentViewNodes.splice(contentIndex, 0, ...childNodes);
+          });
+
+          (current as any).$slot.logicalView = this;
+        }
+      }
+    } else {
+      Reporter.error(16);
+    }
+
+    return observer;
+  }
+
+  public insertVisualChildBefore(visual: IVisual, refNode: INode) {
+    const childObserver = this.childObserver;
+
+    if (childObserver) {
+      const contentNodes = this.childNodes;
+      const contentIndex = contentNodes.indexOf(refNode);
+      const projectedNodes = Array.from(visual.$view.childNodes);
+
+      projectedNodes.forEach((node: any) => {
+        if (node.$isContentProjectionSource) {
+          node.$slot.logicalView = this;
+        }
+      });
+
+      contentNodes.splice(contentIndex, 0, ...projectedNodes);
+      childObserver.notifyChildrenChanged();
+    }
+  }
+
+  public removeVisualChild(visual: IVisual) {
+    const childObserver = this.childObserver;
+
+    if (childObserver) {
+      const contentNodes = this.childNodes;
+      const startIndex = contentNodes.indexOf(visual.$view.firstChild);
+      const endIndex = contentNodes.indexOf(visual.$view.lastChild);
+
+      contentNodes.splice(startIndex, (endIndex - startIndex) + 1);
+      childObserver.notifyChildrenChanged();
+    }
+  }
+
+  public findTargets() { return PLATFORM.emptyArray; }
+  public appendChild(child: INode) {}
+  public insertBefore(refNode: INode): void {}
+  public appendTo(parent: INode): void {}
+  public remove(): void {}
 }
