+++ packages/runtime/src/templating/view.ts
@@ -1,7 +1,7 @@
 import { DI, PLATFORM, Reporter } from '@aurelia/kernel';
 import { IScope } from '../binding/binding-context';
 import { IBindScope } from '../binding/observation';
-import { DOM, IChildObserver, INode, IView } from '../dom';
+import { DOM, INode, IView } from '../dom';
 import { IAttach } from './lifecycle';
 import { IRenderContext } from './render-context';
 import { IRenderSlot } from './render-slot';
@@ -39,16 +39,6 @@ export const noopView: IView = {
 
 export const View = {
   none: noopView,
-  // See below for an explanation of ContentView.
-  // This factory method creates the ContentView depending on whether or not
-  // shadow dom emulation is active on the host element.
-  fromCompiledContent(host: INode, options: { contentOverride?: INode } = PLATFORM.emptyObject): IContentView | null {
-    if (DOM.isUsingSlotEmulation(host)) {
-      return new ContentView(options.contentOverride || host);
-    } else {
-      return null;
-    }
-  },
   // This creates an instance of IView based on an existing node.
   // It's used by the rendering engine to create an instance of IVisual,
   // based on a single component. The rendering engine's createVisualFromComponent
@@ -87,118 +77,3 @@ export const View = {
     };
   }
 };
-
-export interface IContentView extends IView {
-  childObserver?: IChildObserver;
-  insertVisualChildBefore(visual: IVisual, refNode: INode);
-  removeVisualChild(visual: IVisual);
-  attachChildObserver(onChildrenChanged: () => void): IChildObserver;
-}
-
-/*@internal*/
-export interface IContentViewChildObserver extends IChildObserver {
-  notifyChildrenChanged(): void;
-}
-
-// This is the default and only implementation of IContentView.
-// This type of view represents the content that is placed between the opening and closing tags
-// of a custom element. It is only used when ShadowDOM Emulation is required. It enables
-// various interactions between the content, as a logical view, the ShadowDom Emulation,
-// and the RenderSlot.
-// Most if not all of this will go away when we remove the shadow dom emulation and render slot.
-// However, we may need to keep the IChildObserver abstraction for enabling the $children property
-// in different scenarios, depending on what optimizations we make around when/when not to use shadow dom.
-/*@internal*/
-export class ContentView implements IContentView {
-  public firstChild: INode;
-  public lastChild: INode;
-  public childNodes: INode[];
-  public childObserver: IContentViewChildObserver = null;
-
-  constructor(private contentHost: INode) {
-    const childNodes = this.childNodes = Array.from(contentHost.childNodes)
-    this.firstChild = childNodes[0];
-    this.lastChild = childNodes[childNodes.length - 1];
-  }
-
-  public attachChildObserver(onChildrenChanged: () => void): IChildObserver {
-    const contentViewNodes = this.childNodes;
-    let observer = this.childObserver;
-
-    if (!observer) {
-      this.childObserver = observer = {
-        get childNodes() {
-          return contentViewNodes;
-        },
-        disconnect() {
-          onChildrenChanged = null;
-        },
-        notifyChildrenChanged() {
-          if (onChildrenChanged !== null) {
-            onChildrenChanged();
-          }
-        }
-      };
-
-      const workQueue = Array.from(contentViewNodes);
-
-      while(workQueue.length) {
-        const current = workQueue.shift();
-
-        if ((current as any).$isContentProjectionSource) {
-          const contentIndex = contentViewNodes.indexOf(current);
-
-          ((current as any).$slot as IRenderSlot).children.forEach(x => {
-            const childNodes = x.$view.childNodes;
-            childNodes.forEach(x => workQueue.push(x));
-            contentViewNodes.splice(contentIndex, 0, ...childNodes);
-          });
-
-          (current as any).$slot.logicalView = this;
-        }
-      }
-    } else {
-      Reporter.error(16);
-    }
-
-    return observer;
-  }
-
-  public insertVisualChildBefore(visual: IVisual, refNode: INode) {
-    const childObserver = this.childObserver;
-
-    if (childObserver) {
-      const contentNodes = this.childNodes;
-      const contentIndex = contentNodes.indexOf(refNode);
-      const projectedNodes = Array.from(visual.$view.childNodes);
-
-      projectedNodes.forEach((node: any) => {
-        if (node.$isContentProjectionSource) {
-          node.$slot.logicalView = this;
-        }
-      });
-
-      contentNodes.splice(contentIndex, 0, ...projectedNodes);
-      childObserver.notifyChildrenChanged();
-    }
-  }
-
-  public removeVisualChild(visual: IVisual) {
-    const childObserver = this.childObserver;
-
-    if (childObserver) {
-      const contentNodes = this.childNodes;
-      const startIndex = contentNodes.indexOf(visual.$view.firstChild);
-      const endIndex = contentNodes.indexOf(visual.$view.lastChild);
-
-      contentNodes.splice(startIndex, (endIndex - startIndex) + 1);
-      childObserver.notifyChildrenChanged();
-    }
-  }
-
-  public findTargets() { return PLATFORM.emptyArray; }
-  public appendChild(child: INode) {}
-  public insertBefore(refNode: INode): void {}
-  public appendTo(parent: INode): void {}
-  public remove(): void {}
-}
